integration_tests/test-requirements.txt
git+https://github.com/xivo-pbx/xivo-provd-client.git
docker-compose
docker-py
nose
pyhamcrest
psycopg2  # sqlalchemy engine
requests<2.7  # docker-compose 1.3 refuses requests>=2.7
sqlalchemy
